+++ src/main/java/com/aiqin/bms/scmp/api/product/domain/request/sku/supplier/UpdateSkuSupplyUnitReqVo.java
@@ -7,6 +7,7 @@ import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 /**
@@ -38,7 +39,7 @@ public class UpdateSkuSupplyUnitReqVo {
     private String companyName;
 
     @ApiModelProperty("税率")
-    private Long taxRate;
+    private BigDecimal taxRate;
 
     @ApiModelProperty(value = "sku供应商信息")
     @JsonProperty("productSkuSupplyUnits")
