+++ src/main/java/com/aiqin/bms/scmp/api/product/domain/response/sku/ProductSkuCheckoutRespVo.java
@@ -32,13 +32,13 @@ public class ProductSkuCheckoutRespVo extends CommonBean {
     private String settlementMethodName;
 
     @ApiModelProperty("销项税率")
-    private Long outputTaxRate;
+    private BigDecimal outputTaxRate;
 
     @ApiModelProperty("进项税率")
-    private Long inputTaxRate;
+    private BigDecimal inputTaxRate;
 
     @ApiModelProperty("积分系数")
-    private Long integralCoefficient;
+    private BigDecimal integralCoefficient;
 
     @ApiModelProperty("物流费奖励比例")
     private BigDecimal logisticsFeeAwardRatio;
