+++ src/main/java/com/aiqin/bms/scmp/api/purchase/domain/request/OutboundDetailRequest.java
@@ -3,6 +3,8 @@ package com.aiqin.bms.scmp.api.purchase.domain.request;
 import io.swagger.annotations.ApiModelProperty;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 /**
  * <p>
  * ━━━━━━神兽出没━━━━━━
@@ -39,7 +41,7 @@ public class OutboundDetailRequest {
     private String skuCode;
 
     @ApiModelProperty("渠道单价")
-    private Long channelUnitPrice;
+    private BigDecimal channelUnitPrice;
 
     @ApiModelProperty("数量")
     private Long num;
